Refactoring to const.py
